+++ py2max/core.py
@@ -1,4 +1,4 @@
-"""py2max: a pure python script to generate .maxpat patcher files.
+"""py2max: a pure python library to generate .maxpat patcher files.
 
 basic usage:
 
@@ -12,6 +12,9 @@ basic usage:
 
 """
 import json
+import os
+
+from .odb import OBJECT_DEFAULTS
 
 LAYOUT_DEFAULT_PAD = 32.0
 LAYOUT_DEFAULT_BOX_WIDTH = 66.0
@@ -55,6 +58,23 @@ class TextBox(Box):
         super().__init__(id, maxclass, numinlets, numoutlets, outlettype,
                          patching_rect, varname, **kwds)
         self.text = text
+        self._kwds = kwds
+        self._parse(text)
+
+    def _parse(self, text):
+        """Extracts maxclass from text and updates fields from defaults.
+        """
+        maxclass = text.split()[0]
+        try:
+            props = OBJECT_DEFAULTS[maxclass]
+            x1, y1, _, _ = self.__dict__['patching_rect']
+            _, _, w2, h2 = props['patching_rect']
+            props['patching_rect'] = [x1, y1, w2, h2]
+            props.update(self._kwds)
+            self.__dict__.update(props)
+        except:
+            # print(f"warning: no default for {maxclass}")
+            pass
 
 
 class CommentBox(TextBox):
@@ -132,23 +152,30 @@ class IntParam(NumberBox):
         self.minimum = minimum
         self.hint = hint
 
+
+
 class Patchline:
+    """A class for Max patchlines."""
+
     def __init__(self, src: str, src_outlet: int, dst: str, dst_inlet: int, order: int = 0):
         self.source = [src, src_outlet]
         self.destination = [dst, dst_inlet]
         self.order = order
 
     def to_tuple(self):
-        """return a tuple describing the patchline"""
+        """Return a tuple describing the patchline."""
         return (self.source[0], self.source[1], 
                 self.destination[0], self.destination[1], 
                 self.order)
 
     def to_dict(self):
-        """create dict from object"""
+        """create dict from object."""
         return dict(patchline=vars(self))
 
+
 class PositionManager:
+    """Utiltiy class to help with object position calculations."""
+
     def __init__(self, parent):
         self.parent = parent
         self.pad = LAYOUT_DEFAULT_PAD
@@ -157,30 +184,35 @@ class PositionManager:
     
     @property
     def patcher_rect(self):
+        """return rect coordinates of the parent patcher"""
         return self.parent.rect
 
     def above(self, rect):
+        """Return a position above the object"""
         x, y, w, h = rect
         return [x, y - self.box_height, w, h]
 
     def below(self, rect):
+        """Return a position below the object"""
         x, y, w, h = rect
         return [x, y + (self.box_height + h), w, h]
 
     def left(self, rect):
+        """Return a position left of the object"""
         x, y, w, h = rect
         return [x - self.box_width, y, w, h]
 
     def right(self, rect):
+        """Return a position right of the object"""
         x, y, w, h = rect
         return [x + (self.box_width + w), y, w, h]
 
 
 
 class Patcher:
-    """Build Max patchers from the ground up
+    """Core Patcher class describing a Max patchers from the ground up.
     
-    Top level Patcher can be converted to .maxpat file
+    Any Patcher can be converted to a .maxpat file.
     """
 
     def __init__(self, path=None, parent=None, classnamespace=None, 
@@ -253,18 +285,22 @@ class Patcher:
         return dict(patcher=d)
 
     def to_json(self):
+        """cascade convert to json"""
         self.render()
         return json.dumps(self.to_dict(), indent=4)
 
     @property
     def width(self):
+        """width of patcher window."""
         return self.rect[2]
 
     @property
     def height(self):
+        """height of patcher windows."""
         return self.rect[3]
 
     def render(self, reset=False):
+        """cascade convert py2max objects to dicts."""
         if reset or self._reset_on_render:
             self.boxes = []
             self.lines = []
@@ -274,26 +310,31 @@ class Patcher:
         self.lines = [line.to_dict() for line in self._lines]
 
     def saveas(self, path):
+        """save as .maxpat json file"""
+        parent_dir = os.path.dirname(path)
+        os.makedirs(parent_dir, exist_ok=True)
         self.render()
         with open(path, 'w') as f:
             json.dump(self.to_dict(), f, indent=4)
 
     def save(self):
+        """save as json .maxpat file"""
         self.saveas(self._path)
 
     def get_id(self):
+        """helper func to increment object ids"""
         self._id_counter += 1
         return f'obj-{self._id_counter}'
  
     def get_pos(self):
-        """rough auto-layout of objects"""
+        """helper func providing very rough auto-layout of objects"""
         pad = LAYOUT_DEFAULT_PAD  # 32.0
         x_pad = pad
         y_pad = pad
         x_shift =   3 * pad * self._x_layout_counter
         y_shift = 1.5 * pad * self._y_layout_counter
         x = x_pad + x_shift
-        w = LAYOUT_DEFAULT_BOX_WIDTH  # 66.0
+        w = LAYOUT_DEFAULT_BOX_WIDTH   # 66.0
         h = LAYOUT_DEFAULT_BOX_HEIGHT  # 22.0
         self._x_layout_counter += 1
         if x + w + 2 * x_pad > self.width:
@@ -304,6 +345,7 @@ class Patcher:
 
     def get_link_order(self, src_id, dst_id):
         """get order of lines between the same pair of objects"""
+
         if ((src_id, dst_id) == self._last_link):
             self._link_counter += 1
         else:
@@ -313,6 +355,7 @@ class Patcher:
 
     def add_box(self, box, comment=None, comment_pos=None):
         """registers the box and adds it to the patcher"""
+
         self._ids.append(box.id)
         self._objects[box.id] = box
         self._boxes.append(box)
@@ -322,7 +365,8 @@ class Patcher:
         return box
 
     def add_associated_comment(self, rect, comment, comment_pos=None):
-        """addd a comment associated with the object"""
+        """add a comment associated with the object"""
+
         if comment_pos:
             assert comment_pos in ['above', 'below', 'right', 'left']
             patching_rect = getattr(self._pos_mgr, comment_pos)(rect)
@@ -330,12 +374,16 @@ class Patcher:
             patching_rect = self._pos_mgr.above(rect)
         self.add_comment(comment, patching_rect)
 
-    def add_patchline_by_index(self, src_i: int, src_outlet, dst_i, dst_inlet):
+    def add_patchline_by_index(self, src_i, src_outlet, dst_i, dst_inlet):
+        """Patchline creation between two objects using stored indexes"""
+
         src_id = self._ids[src_i]
         dst_id = self._ids[dst_i]
         self.add_patchline(src_id, src_outlet, dst_id, dst_inlet)
 
     def add_patchline(self, src_id, src_outlet, dst_id, dst_inlet):
+        """primary patchline creation method"""
+
         order = self.get_link_order(src_id, dst_id)
         patchline = Patchline(src_id, src_outlet, dst_id, dst_inlet, order)
         self._lines.append(patchline)
@@ -343,12 +391,17 @@ class Patcher:
 
     def add_line(self, src_obj, dst_obj, outlet=0, inlet=0):
         """convenience line adding taking objects with default outlet to inlet"""
+
         return self.add_patchline(src_obj.id, outlet, dst_obj.id, inlet)
 
     def add_textbox(self, text: str, maxclass: str = None, 
                     numinlets: int = None, numoutlets: int = None, outlettype: list[str] = None, 
                     patching_rect: list[float] = None, id: str = None, varname: str = None, 
                     comment: str = None, comment_pos: str = None, **kwds):
+        """Add a generic textbox object to the patcher.
+
+        Looks up default attributes from a dictionary.
+        """
 
         return self.add_box(
             TextBox(
@@ -508,9 +561,6 @@ class Patcher:
             comment_pos
         )
 
-# id: str, maxclass: str, text: str,
-# numinlets: int, numoutlets: int, outlettype: list[str],
-# patching_rect: list[float], varname: str = None, **kwds
 
 class SubPatcher(TextBox):
     """Subpatcher textbox subclass"""
@@ -530,91 +580,3 @@ class SubPatcher(TextBox):
     @property
     def subpatcher(self):
         return self._patcher
-
-
-def test_colors():
-    p = Patcher('out.maxpat')
-    for i in range(54):
-        m = i/54.
-        p.add_textbox('cycle~ 400', bgcolor=[1.0-m, 0.32, 0.0+m, 0.5])
-    p.save()
-
-def test_basic():
-    p = Patcher('out.maxpat')
-    osc1 = p.add_textbox('cycle~ 440')
-    gain = p.add_textbox('gain~')
-    dac = p.add_textbox('ezdac~')
-    p.add_line(osc1, gain)
-    p.add_line(gain, dac)
-    p.save()
-
-def test_subpatch():
-    p = Patcher('out.maxpat')
-    sbox = p.add_subpatcher('p mysub')
-    sp = sbox.subpatcher
-    i = sp.add_textbox('inlet')
-    g = sp.add_textbox('gain~')
-    o = sp.add_textbox('outlet')
-    osc = p.add_textbox('cycle~ 440')
-    dac = p.add_textbox('ezdac~')
-    sp.add_line(i, g)
-    sp.add_line(g, o)
-    p.add_line(osc, sbox)
-    p.add_line(sbox, dac)
-    p.save()
-
-def test_param():
-    p = Patcher('out.maxpat')
-    freq = p.add_floatparam('frequency1', 230, 0, 1000)
-    size = p.add_intparam('size', 341, 0, 1000)
-    p.save()
-
-def test_comment():
-    p = Patcher('out.maxpat')
-    osc1 = p.add_textbox('cycle~ 440', comment="sine osc")
-    fparam = p.add_floatparam('Frequency', 10.2, 0.0, 20.0)
-    p.save()
-
-def test_mix_two_sinusoids():
-    p = Patcher('out.maxpat')
-    
-    # objects
-    freq1 = p.add_floatparam('frequency1', 230, 0, 1000)
-    freq2 = p.add_floatparam('frequency2', 341, 0, 1000)
-    phase = p.add_floatbox()
-    osc1 = p.add_textbox('cycle~')
-    osc2 = p.add_textbox('cycle~')
-    amp1 = p.add_floatbox()
-    amp2 = p.add_floatbox()
-    mul1 = p.add_textbox('*~')
-    mul2 = p.add_textbox('*~')
-    add1 = p.add_textbox('+~')
-    dac = p.add_textbox('ezdac~')
-    scop = p.add_textbox('scope~')
-    scp1 = p.add_intbox()
-    scp2 = p.add_intbox()
-
-    # lines
-    p.add_line(freq1, osc1)
-    p.add_line(osc1, mul1)
-    p.add_line(mul1, add1)
-    p.add_line(amp1, mul1, inlet=1)
-    p.add_line(freq2, osc2)
-    p.add_line(phase, osc2, inlet=1)
-    p.add_line(osc2, mul2)
-    p.add_line(amp2, mul2, inlet=1)
-    p.add_line(mul2, add1, inlet=1)
-    p.add_line(add1, dac)
-    p.add_line(add1, dac, inlet=1)
-    p.add_line(add1, scop)
-    p.add_line(scp1, scop)
-    p.add_line(scp2, scop, inlet=1)
-    p.save()
-
-if __name__ == '__main__':
-    # test_colors()
-    # test_basic()
-    # test_subpatch()
-    # test_param()
-    # test_mix_two_sinusoids()
-    test_comment()