+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 0.1
 
+
+- smarter textbox which uses odb to improve object creation.
+- separate test folder
+- added odb.py in package with a number of default configs of objects
+- converted to package.
+
 - added some notes on graph drawing and layout algorithms
 - comments keyword in box objects + PositionManager for easy documentation
 - comments objects added